refactor: client-browser package with localstorage keystore, polyfills, and browserify bundling
