+++ packages/client-browser/browser-exports.js
@@ -1,0 +1,2 @@
+window.nearApi = require('./lib');
+window.Buffer = Buffer;
