+++ packages/client-browser/src/index.ts
@@ -1,0 +1,3 @@
+import 'error-polyfill';
+
+export * from './key_store';
