+++ packages/client-browser/src/key_store/index.ts
@@ -1,0 +1,1 @@
+export { BrowserLocalStorageKeyStore } from './browser_local_storage_key_store';
