+++ packages/near-api-js/package.json
@@ -11,6 +11,7 @@
     "browser": "lib/browser-index.js",
     "types": "lib/index.d.ts",
     "dependencies": {
+        "@near-js/client-browser": "workspace:*",
         "@near-js/client-core": "workspace:*",
         "@near-js/client-node": "workspace:*",
         "@near-js/providers": "workspace:*",
